requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
google-cloud-bigquery==3.33.0
python-dotenv==1.0.0
pytest==8.2.2
